requirements.txt
Flask>=3.0
Flask-JWT-Extended>=4.6
Flask-Bcrypt>=1.0
Flask-SQLAlchemy>=3.1
PyMySQL>=1.1
python-dotenv>=1.0
